Edit the file structure of the project
