+++ internal/table/table_test.go
@@ -1,4 +1,4 @@
-package main
+package table
 
 import (
 	"reflect"
